# CateMate-Workbench/lib/api-zod/src/generated/types/blueprint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CateMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlueprintChapter } from './blueprintChapter';

export interface Blueprint {
  chapters: BlueprintChapter[];
  /** @nullable */
  verdict?: string | null;
  /** @nullable */
  rawJson?: string | null;
}