# CateMate-Workbench/lib/api-zod/src/generated/types/blueprintChapter.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CateMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlueprintChapterScopeKind } from './blueprintChapterScopeKind';

export interface BlueprintChapter {
  id: string;
  title: string;
  modules: string[];
  metrics?: string[];
  /** @nullable */
  scopeKind?: BlueprintChapterScopeKind;
}