# CateMate-Workbench/lib/api-zod/src/generated/types/dataSourceIngestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CateMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DataSourceIngestInputType } from './dataSourceIngestInputType';

export interface DataSourceIngestInput {
  path: string;
  category: string;
  type: DataSourceIngestInputType;
}