# CateMate-Workbench/lib/api-zod/src/generated/types/planningMode.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CateMate API specification
 * OpenAPI spec version: 0.1.0
 */

export type PlanningMode = typeof PlanningMode[keyof typeof PlanningMode];


export const PlanningMode = {
  v2_solve_loop: 'v2_solve_loop',
  module_selection: 'module_selection',
  ai_direct: 'ai_direct',
} as const;