# CateMate-Workbench/lib/api-zod/src/generated/types/run.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CateMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClarifyingQuestion } from './clarifyingQuestion';
import type { DeliverableSet } from './deliverableSet';
import type { PlanningMode } from './planningMode';
import type { RunStatus } from './runStatus';
import type { SolveProgress } from './solveProgress';
import type { UnderstandingSpec } from './understandingSpec';

export interface Run {
  id: string;
  caseId: string;
  status: RunStatus;
  planningMode: PlanningMode;
  requirementText: string;
  /** @nullable */
  site?: string | null;
  /** @nullable */
  category?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  understanding?: UnderstandingSpec;
  /** @nullable */
  clarifyingQuestions?: ClarifyingQuestion[] | null;
  deliverables?: DeliverableSet;
  solveProgress?: SolveProgress;
}