# CateMate-Workbench/lib/api-zod/src/generated/types/solvePhase.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CateMate API specification
 * OpenAPI spec version: 0.1.0
 */

export type SolvePhase = typeof SolvePhase[keyof typeof SolvePhase];


export const SolvePhase = {
  blueprint: 'blueprint',
  plan: 'plan',
  catalog: 'catalog',
  execute: 'execute',
  verify: 'verify',
  done: 'done',
} as const;